ml-service/requirements.txt
tensorflow==2.13.0
flask==2.3.3
flask-cors==4.0.0
Pillow==10.0.0
numpy==1.24.3
